requirements.txt
flask==3.0.3
requests==2.31.0
gunicorn==22.0.0
google-cloud-secret-manager==2.20.2